backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
timm
Pillow
python-multipart
numpy